IGNITE-7040 Web Console: Refactored list of registered users. (cherry picked from commit 31c60bd)
