+++ modules/web-console/frontend/app/components/list-of-registered-users/categories.js
