+++ modules/web-console/frontend/app/components/list-of-registered-users/template.tpl.pug
@@ -1,0 +1,65 @@
+//-
+    Licensed to the Apache Software Foundation (ASF) under one or more
+    contributor license agreements.  See the NOTICE file distributed with
+    this work for additional information regarding copyright ownership.
+    The ASF licenses this file to You under the Apache License, Version 2.0
+    (the "License"); you may not use this file except in compliance with
+    the License.  You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+include /app/helpers/jade/mixins
+
+ul.tabs.tabs--blue
+    li(role='presentation' ng-class='{ active: $ctrl.groupBy === "user" }') 
+        a(ng-click='$ctrl.groupByUser()') 
+            span Users
+            span.badge.badge--blue(ng-hide='$ctrl.groupBy === "user"')
+                | {{ $ctrl.gridOptions.data.length }}
+            span.badge.badge--blue(ng-show='$ctrl.groupBy === "user"')
+                | {{ $ctrl.filteredRows.length }}
+    li(role='presentation' ng-class='{ active: $ctrl.groupBy === "company" }')
+        a(ng-click='$ctrl.groupByCompany()') 
+            span Companies
+            span.badge.badge--blue {{ $ctrl.companies.length }}
+    li(role='presentation' ng-class='{ active: $ctrl.groupBy === "country" }')
+        a(ng-click='$ctrl.groupByCountry()')
+            span Countries
+            span.badge.badge--blue {{ $ctrl.countries.length }}
+
+.panel--ignite
+    .panel-heading.ui-grid-settings
+        .panel-title
+            +ignite-form-field-bsdropdown({
+                label: 'Actions',
+                model: '$ctrl.action',
+                name: 'action',
+                disabled: '!$ctrl.selected.length',
+                required: false,
+                options: '$ctrl.actionOptions'
+            })
+            grid-export(grid-api='$ctrl.gridApi')
+            form.ui-grid-settings-dateperiod(name=form novalidate)
+                -var form = 'admin'
+                +ignite-form-field-datepicker('Period: from', '$ctrl.params.startDate', '"startDate"', null, '$ctrl.params.endDate')
+                +ignite-form-field-datepicker('to', '$ctrl.params.endDate', '"endDate"', '$ctrl.params.startDate', null)
+            form.ui-grid-settings-filter
+                -var form = 'admin'
+                +ignite-form-field-text('Exclude:', '$ctrl.params.companiesExclude', '"exclude"', false, false, 'Exclude by company name...')
+
+            .ui-grid-settings--heading(ng-hide='$ctrl.selected.length')
+                span(ng-if='$ctrl.groupBy === "user"') List of registered users
+                span(ng-if='$ctrl.groupBy === "company"') List of registered companies
+                span(ng-if='$ctrl.groupBy === "country"') List of registered countries
+                grid-column-selector(grid-api='$ctrl.gridApi')
+            .panel-selected(ng-show='$ctrl.selected.length')
+                grid-item-selected(grid-api='$ctrl.gridApi')
+
+    .panel-collapse
+        .grid.ui-grid--ignite.ui-grid-disabled-group-selection(ui-grid='$ctrl.gridOptions' ui-grid-resize-columns ui-grid-selection ui-grid-exporter ui-grid-pinning ui-grid-grouping ui-grid-hovering)
